Converted label_batch to long

+++ src/dataloader.py
@@ -9,7 +9,7 @@ from dataset import CustomDataset, ImageDataset
 class TrainViewDataLoader(DataLoader):
     def __iter__(self):
         data_batch = torch.Tensor()
-        label_batch = torch.Tensor()
+        label_batch = torch.Tensor().long()
         for idx in self.sampler:
             data, labels = self.dataset[idx]
             data_batch = torch.cat([data_batch, data])
@@ -35,7 +35,7 @@ class TestViewDataLoader(DataLoader):
     def __iter__(self):
         for idx in self.sampler:
             batch = torch.Tensor()
-            batch2 = torch.Tensor()
+            batch2 = torch.Tensor().long()
             data, labels = self.dataset[idx]
             index = torch.tensor([0])
             yield torch.cat([batch, data]), torch.index_select(torch.cat([batch2, labels]), 0, index)
